This is a big commit - [x] Added some stuff, fixed some stuff, removed some stuff: the normal
