--- .hound.yml
@@ -1,1 +1,0 @@
-esversion: 6
