+++ .nojekyll
@@ -1,1 +1,6 @@
 DO NOT DELETE THIS FILE- THANKS
+
+
+If you really love jekyll inside;
+I don't use jekyll,
+I'd better hyde
